+++ app/routes/__init__.py
@@ -6,8 +6,3 @@ api_v1 = Blueprint('api_v1', __name__, url_prefix='')
 
 # API version 1.
 from .v1 import confirmed, deaths, recovered, all
-
-# Redirect to project page on index.
-@app.route('/')
-def index():
-    return redirect('https://github.com/ExpDev07/coronavirus-tracker-api', 302)
